File renaming

+++ cuneiform_src/Kern/hhh/puma.h
@@ -222,7 +222,7 @@ bool LPUMA_GetPreserveLineBreaks(void);
 
 uint32_t LCED_DeletePage(Handle hEdPage);
 #define DEC_FUN(a,b,c) typedef a (*FN##b)c; PUMA_FUNC a b c; a L##b c;
-PUMA_FUNC bool PUMA_Init(uint16_t wHeightCode, Handle hStorage);
+PUMA_FUNC bool PUMA_Init(uint wHeightCode, Handle hStorage);
 DEC_FUN(bool, PUMA_Done,())
 DEC_FUN(uint32_t, PUMA_GetReturnCode,())
 DEC_FUN(char *, PUMA_GetReturnString,(uint32_t dwError))
